third_party/blink/web_tests/http/tests/security/contentSecurityPolicy/sandbox-allow-scripts-in-http-header-expected.txt

CONSOLE MESSAGE: Script executed in iframe.
CONSOLE MESSAGE: PASS: Iframe was in a unique origin
